+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
  * The base of deviation of local variable abuse has been changed from the average number of rows to the first number of rows.
  * AssignOp has been added to the assignment decision in addition to Assign.
-
+ * Add filename to report json.
 
 ## v0.0.2 (2025-03-19)